+++ Leomard/Services/PostService.swift
@@ -24,7 +24,8 @@ class PostService: Service {
                         var responses = try self.decode(type: GetPostsResponse.self, data: data)
                         responses.posts.forEach { postView in
                             if postView.post.nsfw && !UserPreferences.getInstance.showNsfw ||
-                                postView.post.nsfw && !UserPreferences.getInstance.showNsfwInFeed {
+                                postView.post.nsfw && !UserPreferences.getInstance.showNsfwInFeed ||
+                                postView.read && UserPreferences.getInstance.hideReadPosts {
                                 responses.posts = responses.posts.filter { $0 != postView}
                             }
                         }
@@ -62,7 +63,8 @@ class PostService: Service {
                     do {
                         var responses = try self.decode(type: GetPostsResponse.self, data: data)
                         responses.posts.forEach { postView in
-                            if postView.post.nsfw && !UserPreferences.getInstance.showNsfw  {
+                            if postView.post.nsfw && !UserPreferences.getInstance.showNsfw ||
+                                postView.read && UserPreferences.getInstance.hideReadPosts{
                                 responses.posts = responses.posts.filter { $0 != postView}
                             }
                         }