+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Color coding to subcomments
 - Comments can now be marked as read manually, on post view, or vote
 - "Show NSFW Content in Feed" toggle
+- "Hide Read Posts" toggle
 
 ### Changes